replace storage layer with Cassandra

+++ models/student.go
@@ -1,8 +1,8 @@
 package models
 
 type Student struct {
-	ID    string  `json:"id"`
-	Name  string  `json:"name"`
-	Email string  `json:"email"`
-	GPA   float64 `json:"gpa"`
+	ID    string `json:"id"`
+	Name  string `json:"name"`
+	Email string `json:"email"`
+	GPA   int    `json:"gpa"`
 }
